+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # QOwnNotes Changelog
 
+## 0.61
+- only todo items (not calendar items) are now shown in your todo list
+
 ## 0.60
 - if the `Ctrl` key is pressed in the note text edit to click a link the mouse cursor is now a pointing hand
 - you are now able to view your **ownCloud todo lists** (ownCloud tasks) in QOwnNotes (readonly so far)